requirements.txt
Pillow==9.2.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
qrcode==7.3.1
